Move fetchData to its own file to make mocking easier
